# openapi_server/models/work_log.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class WorkLog(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, _date=None, description=None):  # noqa: E501
        """WorkLog - a model defined in OpenAPI

        :param id: The id of this WorkLog.  # noqa: E501
        :type id: int
        :param _date: The _date of this WorkLog.  # noqa: E501
        :type _date: date
        :param description: The description of this WorkLog.  # noqa: E501
        :type description: str
        """
        self.openapi_types = {
            'id': int,
            '_date': date,
            'description': str
        }

        self.attribute_map = {
            'id': 'id',
            '_date': 'date',
            'description': 'description'
        }

        self._id = id
        self.__date = _date
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'WorkLog':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The WorkLog of this WorkLog.  # noqa: E501
        :rtype: WorkLog
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this WorkLog.


        :return: The id of this WorkLog.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this WorkLog.


        :param id: The id of this WorkLog.
        :type id: int
        """

        self._id = id

    @property
    def _date(self) -> date:
        """Gets the _date of this WorkLog.


        :return: The _date of this WorkLog.
        :rtype: date
        """
        return self.__date

    @_date.setter
    def _date(self, _date: date):
        """Sets the _date of this WorkLog.


        :param _date: The _date of this WorkLog.
        :type _date: date
        """

        self.__date = _date

    @property
    def description(self) -> str:
        """Gets the description of this WorkLog.


        :return: The description of this WorkLog.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this WorkLog.


        :param description: The description of this WorkLog.
        :type description: str
        """

        self._description = description